Fix XAML selection behavior resolution
